geneset overhaul and download/upload dialogs complete
