+++ backend/server/__init__.py
@@ -2,7 +2,7 @@ import logging
 import sys
 from backend.common.utils.utils import import_plugins
 
-__version__ = "2.7.6"
+__version__ = "2.8.0"
 display_version = "cellxgene v" + __version__
 
 try:
