+++ client/src/components/framework/container.js
@@ -1,5 +1,4 @@
 import React from "react";
-
 function Container(props) {
   const { children } = props;
   return (
