+++ client/src/components/geneExpression/index.js
@@ -1,12 +1,11 @@
 import React from "react";
 import { connect } from "react-redux";
-import { Icon, Collapse, AnchorButton, Tooltip, Position, MenuItem, Dialog } from "@blueprintjs/core";
+import { AnchorButton, Tooltip, Position, MenuItem, Dialog } from "@blueprintjs/core";
 import { Select } from "@blueprintjs/select";
 import ParameterInput from "../menubar/parameterinput";
 import GeneSet from "./geneSet";
 import { GenesetHotkeys } from "../hotkeys";
 import actions from "../../actions";
-import Truncate from "../util/truncate"
 import * as globals from "../../globals";
 import AnnoDialog from "../annoDialog";
 import LabelInput from "../labelInput";
@@ -29,7 +28,6 @@ import QuickGene from "./quickGene";
     annoMatrix: state.annoMatrix,
     reembedParams: state.reembedParameters,
     userLoggedIn: state.controls.userInfo ? true : false,
-    outputController: state.outputController,
     cxgMode: state.controls.cxgMode
   };
 })
@@ -42,10 +40,6 @@ class GeneExpression extends React.Component {
       newNameText: "",
       nameBeingEdited: "",
       varMetadata: "",
-      fileName: "Choose file...",
-      fileName2: "Choose file...",
-      uploadMetadataOpen: false,
-      uploadMetadataOpen2: false
     };
   }
 
@@ -89,120 +83,8 @@ class GeneExpression extends React.Component {
     dispatch({type: "track set", group: newName, set: null})   
     this.disableEditNameMode()
   }
-  setupFileInput = () => {
-    const context = this;
-    const { dispatch } = this.props;
-
-    function uploadDealcsv () {};
-    uploadDealcsv.prototype.getCsv = function(e) {
-      let input = document.getElementById('dealCsv2');
-      input.addEventListener('change', function() {
-        if (this.files && this.files[0]) {
-
-            var myFile = this.files[0];
-            var reader = new FileReader();
-            context.setState({...context.state,fileName: myFile.name})
-            reader.addEventListener('load', function (e) {
-                
-                let csvdata = e.target.result; 
-                parseCsv.getParsecsvdata(csvdata); // calling function for parse csv data 
-                context.resetFileState();
-            });
-            
-            reader.readAsBinaryString(myFile);
-        }
-      });
-    }
-    uploadDealcsv.prototype.getParsecsvdata = function(data) {
-      const genesets = {};
-
-      let newLinebrk = data.split("\n");
-      if (newLinebrk.at(-1)===""){
-        newLinebrk = newLinebrk.slice(0,-1)
-      }
-
-      for(let i = 0; i < newLinebrk.length; i++) {
-        const y = newLinebrk[i].split(/,(?=(?:(?:[^"]*"){2})*[^"]*$)/)
-        const x = [];
-        y.forEach((item)=>{
-          if (item.startsWith("\"") && item.endsWith("\"")){
-            x.push(item.substring(1,item.length-1).split("\r").at(0))
-          } else {
-            x.push(item.split("\r").at(0))
-          }
-        })
-        if (x[0] === "gene_set_description" && x[1] === "gene_set_name"){
-          continue;
-        }
-        const suffix = x[2]==="True" ? "" : "";
-        if (`${x[0]}${suffix}` in genesets) {
-          genesets[`${x[0]}${suffix}`][x[1]] = x.slice(3)
-        } else {
-          genesets[`${x[0]}${suffix}`]={}
-          genesets[`${x[0]}${suffix}`][x[1]] = x.slice(3)
-        }
-      }
-      for (const key1 in genesets) {
-        for (const key2 in genesets[key1]) {
-          dispatch({
-            type: "geneset: create",
-            genesetName: key2,
-            genesetDescription: key1,
-          });
-          dispatch(actions.genesetAddGenes(key1, key2, genesets[key1][key2]));  
-        }
-      }
-  
-
-    }
-    var parseCsv = new uploadDealcsv();
-    parseCsv.getCsv();
-  }
-
-  resetFileState = () => {
-    this.setState({...this.state, 
-      fileName: "Choose file...", 
-      uploadMetadataOpen: false,
-    })
-  }
-
-  setupFileInput2 = () => {
-    const { dispatch, annoMatrix } = this.props;
-    const context = this;
-    function uploadDealcsv () {};
-    uploadDealcsv.prototype.getCsv = function(e) {
-      let input = document.getElementById('dealCsv3');
-      input.addEventListener('change', function() {
-        if (this.files && this.files[0]) {
-
-            var myFile = this.files[0];
-            const formData = new FormData();
-            formData.append("file",myFile);
-            fetch(`${globals.API.prefix}${globals.API.version}uploadVarMetadata`, {method: "POST", body: formData}).then((res)=>{
-              res.json().then((schema)=>{
-                annoMatrix.updateSchema(schema.schema)
-                dispatch({type: "refresh var metadata"})
-                context.resetFileState2();
-              })
-
-              
-            });            
-        }
-      });
-    }
-    var parseCsv = new uploadDealcsv();
-    parseCsv.getCsv();
-  }
-
-  resetFileState2 = () => {
-    this.setState({...this.state, 
-      fileName2: "Choose file...", 
-      uploadMetadataOpen2: false,
-    })
-  }
 
   renderGeneSets = (isGenes=false) => {
-    /*
     if (isGenes) {
       const { allGenes, rightWidth, cxgMode } = this.props;
       const name = cxgMode==="OBS" ? `All genes` : `All cells`
@@ -214,11 +96,13 @@ class GeneExpression extends React.Component {
           setName={name}
           allGenes
           rightWidth={rightWidth}
+          setMode="genes"
+          genesetDescription={""}
         />
       );  
     }
-    */
-    const { dispatch, genesets, rightWidth, cxgMode } = this.props;    
+    
+    const { dispatch, genesets, rightWidth } = this.props;    
     
     const nogroups = [];
     if ("" in genesets) {
@@ -323,119 +207,13 @@ class GeneExpression extends React.Component {
         sets[key]        
       )
     }
-    /*<div key={key} style={style}>
-        <div style={{
-          display: "flex",
-          backgroundColor: "#F0F0F0",
-        }}>
-        <AnchorButton
-          onClick={() => {
-            this.setState({ 
-              [groupName]: !(this.state[groupName]??false)
-            });
-          }}
-          text={<Truncate><span>{groupName}</span></Truncate>}
-          fill
-          minimal
-          rightIcon={(this.state[groupName]??false) ? "chevron-down" : "chevron-right"} small
-        />  
-        <Tooltip
-        content={`Edit ${cOrG} set group name`}
-        position="top"
-        hoverOpenDelay={globals.tooltipHoverOpenDelay}            
-        >             
-        <AnchorButton
-          icon={<Icon icon="edit" iconSize={10} />}     
-          minimal
-          style={{
-            cursor: "pointer",
-          }}
-          onClick={(e) => this.activateEditNameMode(e,groupName)}
-          />  
-        </Tooltip>                 
-        <Tooltip
-        content={`Delete ${cOrG} set group`}
-        position="top"
-        hoverOpenDelay={globals.tooltipHoverOpenDelay}            
-        >
-        <AnchorButton
-          icon={<Icon icon="trash" iconSize={10} />}     
-          minimal
-          intent="danger"
-          style={{
-            cursor: "pointer",
-          }}
-          onClick={() => {
-            dispatch(actions.genesetDeleteGroup(key))
-          }}
-        />    
-        </Tooltip>                        
-      </div>         
-      <Collapse isOpen={this.state[groupName]??false}>
-        {sets[key]}
-      </Collapse>
-    </div>*/
 
     const els2 = [];
     for ( const key in sets2 ){
       els2.push(
         sets2[key]
       )
-    }   
-    /*<div key={key} style={style}>
-          <div style={{
-            display: "flex",
-            backgroundColor: "#F0F0F0",
-          }}>
-          <AnchorButton
-            onClick={() => {
-              this.setState({ 
-                [groupName]: !(this.state[groupName]??false)
-              });
-            }}
-            text={<Truncate><span>{groupName.split('//;;//').at(0)}</span></Truncate>}
-            fill
-            minimal
-            rightIcon={(this.state[groupName]??false) ? "chevron-down" : "chevron-right"} small
-          />  
-          <Tooltip
-          content={`Edit ${cOrG} set group name`}
-          position="top"
-          hoverOpenDelay={globals.tooltipHoverOpenDelay}            
-          >             
-          <AnchorButton
-            icon={<Icon icon="edit" iconSize={10} />}     
-            minimal
-            style={{
-              cursor: "pointer",
-            }}
-            onClick={(e) => this.activateEditNameMode(e,groupName)}
-            />  
-          </Tooltip>                 
-          <Tooltip
-          content={`Delete ${cOrG} set group`}
-          position="top"
-          hoverOpenDelay={globals.tooltipHoverOpenDelay}            
-          >
-          <AnchorButton
-            icon={<Icon icon="trash" iconSize={10} />}     
-            minimal
-            intent="danger"
-            style={{
-              cursor: "pointer",
-            }}
-            onClick={() => {
-              dispatch(actions.genesetDeleteGroup(key))
-              dispatch(actions.requestDiffDelete(key))
-            }}
-          />    
-          </Tooltip>                        
-        </div>         
-        <Collapse isOpen={this.state[groupName]??false}>
-          {sets2[key]}
-        </Collapse>
-      </div>    
-    */
+    }
     return [nogroups, els, els2];      
   };
   
@@ -458,126 +236,36 @@ class GeneExpression extends React.Component {
   };
 
   render() {
-    const { dispatch, genesets, annoMatrix, userLoggedIn, var_keys, outputController, cxgMode, rightWidth } = this.props;
-    const { geneSetsExpanded, isEditingSetName, newNameText, nameBeingEdited, varMetadata, uploadMetadataOpen, fileName, uploadMetadataOpen2, fileName2, preferencesDialogOpen } = this.state;
+    const { dispatch, genesets, annoMatrix, userLoggedIn, var_keys, cxgMode, rightWidth } = this.props;
+    const { isEditingSetName, newNameText, nameBeingEdited, varMetadata, preferencesDialogOpen } = this.state;
     const [nogroupElements,genesetElements,diffExpElements]=this.renderGeneSets();
-    const saveLoading = !!outputController?.pendingFetch;
     const cOrG = cxgMode === "OBS" ? "gene" : "cell";
     return (
-      <div>
+      <div onDrop={(e)=>{
+        dispatch({type: "clear gene selection"})
+        const name = e.dataTransfer.getData("text");   
+        const setgroup = name.split("@@").at(0)
+        const setname = name.split("@@").at(1)             
+        if (!name.includes("@@@")) {
+          dispatch({
+            type: "geneset: update",
+            genesetDescription: setgroup,
+            genesetName: setname,
+            update: {
+              genesetName: setname,
+              genesetDescription: "",
+            },
+            isDragging: true
+          });
+          dispatch({type: "track set", group: "", set: setname})   
+          e.stopPropagation();      
+        }
+      }}>
        {userLoggedIn ?  <GenesetHotkeys
           dispatch={dispatch}
           genesets={genesets}
         /> : null}
-        <div style={{marginBottom: "20px",
-              display: "flex",
-              justifyContent: "left",
-              columnGap: "10px"}}>     
-          <Tooltip
-          content="Save gene metadata to a `.txt` file."
-          position="bottom"
-          hoverOpenDelay={globals.tooltipHoverOpenDelay}
-        >                                              
-          <AnchorButton
-              type="button"
-              loading={saveLoading}
-              icon="floppy-disk"
-              onClick={() => {
-                dispatch(actions.downloadVarMetadata())
-              }}
-            /> 
-          </Tooltip>           
-          <Tooltip
-          content="Upload gene metadata from a tab-delimited .txt file."
-          position="bottom"
-          hoverOpenDelay={globals.tooltipHoverOpenDelay}
-        >                                              
-          <AnchorButton
-              type="button"
-              icon="upload"
-              disabled={!userLoggedIn}
-              onClick={() => {
-                this.setState({...this.state,uploadMetadataOpen2: true})
-              }}
-            /> 
-          </Tooltip>                 
-            <Dialog
-              title="Upload gene metadata file (tab-delimited, .txt)."
-              isOpen={uploadMetadataOpen2}
-              onOpened={()=>this.setupFileInput2()}
-              onClose={()=>{
-                this.resetFileState2();
-                }
-              }
-            >
-              <div style={{
-                display: "flex",
-                flexDirection: "column",
-                paddingLeft: "10px",
-                paddingTop: "10px",
-                width: "90%",
-                margin: "0 auto"
-              }}>
-                <label className="bp3-file-input">
-                  <input type="file" id="dealCsv3"/>
-                  <span className="bp3-file-upload-input">{fileName2}</span>
-                </label>           
-              </div>                                
-            </Dialog> 
 
-
-            {/*this is geneset save/upload*/}           
-            {userLoggedIn && <Tooltip
-              content={`Save ${cOrG} sets a \`.csv\` file.`}
-              position="bottom"
-              hoverOpenDelay={globals.tooltipHoverOpenDelay}
-            >                                              
-              <AnchorButton
-                  type="button"
-                  icon="floppy-disk"
-                  onClick={() => {
-                    this.handleSaveGenedata()
-                  }}
-                /> 
-              </Tooltip> }
-          <Tooltip
-          content={`Upload ${cOrG} sets from a \`.csv\`, comma-delimited file.`}
-          position="bottom"
-          hoverOpenDelay={globals.tooltipHoverOpenDelay}
-        >                                              
-          <AnchorButton
-              type="button"
-              icon="upload"
-              disabled={!userLoggedIn}
-              onClick={() => {
-                this.setState({...this.state,uploadMetadataOpen: true})
-              }}
-            /> 
-          </Tooltip>                 
-            <Dialog
-              title={`Upload ${cOrG} sets file (comma-delimited .csv)`}
-              isOpen={uploadMetadataOpen}
-              onOpened={()=>this.setupFileInput()}
-              onClose={()=>{
-                this.resetFileState();
-                }
-              }
-            >
-              <div style={{
-                display: "flex",
-                flexDirection: "column",
-                paddingLeft: "10px",
-                paddingTop: "10px",
-                width: "90%",
-                margin: "0 auto"
-              }}>
-                <label className="bp3-file-input">
-                  <input type="file" id="dealCsv2"/>
-                  <span className="bp3-file-upload-input">{fileName}</span>
-                </label>           
-              </div>                                
-            </Dialog>                       
-        </div>
         <div style={{
               display: "flex",
               justifyContent: "left",
@@ -695,6 +383,7 @@ class GeneExpression extends React.Component {
                 setMode="genes"
             />}
             </div>
+            {this.renderGeneSets(true)}
             {nogroupElements}                     
             {genesetElements}
             {(diffExpElements.length > 0) && <div style={{paddingBottom: "5px",paddingTop: "5px"}}>
