+++ client/src/index.js
@@ -11,6 +11,12 @@ import store from "./reducers";
 
 FocusStyleManager.onlyShowFocusOnTabs();
 
+document.addEventListener('keydown', (e) => {
+  if (e.code === "KeyA" && e.metaKey) {
+    e.preventDefault();
+  } 
+});
+
 ReactDOM.render(
   <Provider store={store}>
     <HotkeysProvider>
